move utility GNs to own folder
