+++ microscopynodes/blender_objects/surface.py
@@ -3,7 +3,7 @@ import bpy
 from .base import MeshChannelObject
 from ..handle_blender_structs.node_handling import get_socket, group_input_output_for_socket, new_socket, set_modifier_input_socket
 from ..handle_blender_structs.min_keys import min_keys
-from ..min_nodes.geo_nodes.import_microscopy_volume import import_microscopy_volume_node_group
+from ..min_nodes.geo_nodes.utilities.import_microscopy_volume import import_microscopy_volume_node_group
 from ..min_nodes.shader_nodes import set_color_ramp_from_ch
