+++ microscopynodes/min_nodes/geo_nodes/__init__.py
@@ -1,19 +1,12 @@
 from .nodeCrosshatch import crosshatch_node_group
-from .import_microscopy_meshes import import_microscopy_meshes_node_group
-from .import_microscopy_volume import import_microscopy_volume_node_group
-from .nodeMicroscopyGridToPoints import microscopy_grid_to_points_node_group
 from .nodeHolderBundleInputs import holder_bundle_inputs_node_group
 from .annotation import grid_verts_node_group, scale_node_group, scalebox_node_group
 from . import annotation
 from . import combine_channels
 from . import masking
 from . import ops
+from . import utilities
 import bpy
-
-
-ADD_MENU_NODE_GROUPS = [
-    "Microscopy Grid to Points",
-]
 
 
 class MIN_MT_GEOMETRY_NODES_ADD(bpy.types.Menu):
@@ -42,12 +35,11 @@ class MIN_MT_GEOMETRY_NODES_ADD(bpy.types.Menu):
             text="LUTs",
             icon="COLOR",
         )
-        for node_name in ADD_MENU_NODE_GROUPS:
-            operator = layout.operator(
-                "microscopynodes.add_geometry_node_group",
-                text=node_name,
-            )
-            operator.node_name = node_name
+        layout.menu(
+            "MIN_MT_UTILITIES_ADD",
+            text="Utilities",
+            icon="TOOL_SETTINGS",
+        )
 
 
 def MIN_add_geometry_node_menu(self, context):
@@ -66,17 +58,16 @@ def MIN_add_geometry_node_menu(self, context):
 
 NODE_GROUPS = {
     "crosshatch": crosshatch_node_group,
-    "Import Microscopy Meshes": import_microscopy_meshes_node_group,
-    "Import Microscopy Volume": import_microscopy_volume_node_group,
-    "Microscopy Grid to Points": microscopy_grid_to_points_node_group,
+    "Holder Bundle Inputs": holder_bundle_inputs_node_group,
     **annotation.NODE_GROUPS,
     **combine_channels.NODE_GROUPS,
     **masking.NODE_GROUPS,
+    **utilities.NODE_GROUPS,
 }
 
 CLASSES = [
     MIN_MT_GEOMETRY_NODES_ADD,
-] + annotation.CLASSES + combine_channels.CLASSES + masking.CLASSES + ops.CLASSES
+] + annotation.CLASSES + combine_channels.CLASSES + masking.CLASSES + utilities.CLASSES + ops.CLASSES
 
 
 def geometry_node_group(name):
