+++ microscopynodes/min_nodes/geo_nodes/utilities/import_microscopy_volume.py
@@ -10,7 +10,7 @@ from nodebpy.types import (
     InputString,
 )
 
-from .nodeHolderBundleInputs import HolderBundleInputs
+from ..nodeHolderBundleInputs import HolderBundleInputs
 
 
 GROUP_NAME = "Import Microscopy Volume"
